fetchParkingLotsWithDistance 사용x
