+++ src/components/ParkingDetail.jsx
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from "react";
 import styles from "./ParkingDetail.module.css";
+import { getDistanceFromLatLonInKm } from "../utils/geo.js";
 
 const ParkingDetail = ({ lot, onClose, onBackToList }) => {
   const [isFavorite, setIsFavorite] = useState(false);
@@ -43,24 +44,6 @@ const ParkingDetail = ({ lot, onClose, onBackToList }) => {
     slotCount,
     availableSpots,
   } = lot;
-
-  function getDistanceFromLatLonInKm(lat1, lon1, lat2, lon2) {
-    const R = 6371;
-    const dLat = deg2rad(lat2 - lat1);
-    const dLon = deg2rad(lon2 - lon1);
-    const a =
-      Math.sin(dLat / 2) * Math.sin(dLat / 2) +
-      Math.cos(deg2rad(lat1)) *
-        Math.cos(deg2rad(lat2)) *
-        Math.sin(dLon / 2) *
-        Math.sin(dLon / 2);
-    const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-    return R * c;
-  }
-
-  function deg2rad(deg) {
-    return deg * (Math.PI / 180);
-  }
 
   const handleHeartClick = () => {
     setIsFavorite(!isFavorite);
